fix: fix bug where dark theme wasn't being applied
Tailwind was configured to only look inside of the src folder to build the css bundle with the used classes.
Since I moved dark to index.html which is outside src it wasn't including the dark root variables.
Didn't notice it earlier cause I'd have to save the index.css file or checkout to another branch to create the issue.

Not the best solution but I don't want to deal with the light/dark theming yet so this works as a temporary solution.